Variable Type in ConfigFile loop
This is a bizarre one.  One one Primary Site server, when the script would Invoke-Expression on $Data[1] where $Data[0] was "LogFile", it caused Notepad.exe to open the $Data[1] (i.e. notepad opened the log file) instead of setting the variable.  This code was derived from https://stackoverflow.com/questions/23042428/how-do-i-dynamically-set-the-type-of-powershell-variables to force setting the variable [type] for numbers and strings... or rather anything that is not an arrary or hashtable.  My testing proved out every variable type (switch/bool), string, int, array, and hashtable.  I added an additional verbose messaging to show the attempt and result.