Config file value is numeric
The regex to test for "is numeric" is actually testing for the existence of any number.  In the case of LogFile=\\server01\share\path\file.ext, the \d evaluates to true.  Then the value -as [int] doesn't throw an error (by design?) but leave the variable LogFile as null/empty.